backend/requirements.txt
# Core Django and API
Django>=5.2.0
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.5.0
django-cors-headers>=4.8.0
python-dotenv>=1.0.0

# File handling
Pillow>=9.5.0

# Database (for production)
psycopg2-binary>=2.9.0

# JWT dependencies
PyJWT>=2.10.0

# Development dependencies
pytest-django>=4.5.0
pytest-cov>=4.0.0
factory-boy>=3.2.0
black>=23.0.0
flake8>=6.0.0